Update sh_class.lua
I Seriously don't understand what it's supposed to do. The only thing it's doing is preventing loading perk choices on subclasses on player init, Everything works as intended when I remove this. The only issues this might cause is if class_choices.txt doesn't exist, but line 585 ensures that it will be fine. I really don't know