Fix for Matriarch Womb's special headcrab relationships
The headcrab spawned by Matriarch Womb currently has a bug where the relationship flags are not set properly for it, causing player minions to be able to target it, which does not end up doing anything since minions cannot damage each other.
This fix sets up the relationship flags properly.