Fixed chaos spawn times (#141)
* Fixed chaos spawn times

I changed how it checks the difficulty because the order it runs the files made it not possible to check the difficulty the way it was before. This should make chaos what it was supposed to be. Also changed the number to 7 because i believe the numbers got changed so a difficulty wasn't set to 0

* Changing it back to adjust it another way

* Redoing the same change

changing the file load order to make the other way possible could cause complications