Enemy number balance (#143)
* Set numbers more accordingly 

I plan on keeping the difficulty variable for enemy count available but mostly unused to keep balancing each difficulty much easier. With a consistent amount of enemies it will be easier to adjust health and damage for them on a difficulty basis. I also removed a "coefficient" variable that isn't used anywhere else

* Regulating enemy count multipliers

This change is to set all but swarm (which intentionally has higher numbers) to a 1x enemy count multiplier. This change is necessary due to the other change I just made that creates a reasonable and now consistent number of enemies per wave across all difficulties.