China Lake, P2k and Fixes (Oh my!) (#173)
* China Lake assets and initial Lua file

Currently a copy of the M32's Lua, will work on making it unique Soon™.

* Removed BO3 MAX-GL model

Oops

* BO1 China Lake: Part 2 (plus some fixes)

Include the model for the China Lake (because I forgor :skull:). Also includes a potential fix for the P2000 as well as the third-person hold for the new M79.

Testing pending.

* Added China Lake to shop. Fixed M202 Reload SFX.

M202 reload also got an ever-so-slight buff in the process, reloading 4% faster.

* Completed China Lake. String and Kill Icon stuff.

I believe the China Lake is ready for prime time!

Also changed the LAW's kill icon, added a material for the FAL's kill icon, and tweaked some descriptions in the shop for a few Demolitionist weapons.

* Update shared.lua

* Update China Lake & M79 models

Removed PaP references from the China Lake model. Also gave M79 a Physics model.

* Got ahead of myself a bit

Removed new M32 sounds I accidentally added to the main branch

* Set China Lake weight to 6, lowered LAW weight to 9

LAW can now be carried alongside a China Lake.