Weapon selling and others
Reverted selling items to 25% and also included auto selling items and other stuff that I forgot to put in the previous pull request.

Also added a lot more checks for functionality to the Medkit. More info on it in the patch notes so that way I don't have to write it here.

Added patch notes for both github and the in game menu that nobody probably will ever see or read.