Extra fixes
The timer is for in some cases where Mind ui gets shown over Armor ui when swapping subclasses.

Added an extra check so that way when swapping classes you don't drop your Medkit. Also made Medkit more forgiving to use.

Moved the patch notes info to the English localization file so that way we can employ CrystalFire for the next few days through torturous translating slave labor. Even though the patch notes are buried in the F2 menu and nobody will read it, it'll be worth the weight.